requirements.txt
cryptography==42.0.8
numpy==1.23.5
pandas==2.2.2
Pillow==9.3.0
Pillow==10.4.0
plotly==5.22.0
scikit_learn==1.5.1
selenium==4.22.0
SQLAlchemy==2.0.31
streamlit==1.36.0
